+++ src/main/java/mod/gottsch/forge/dungeons2/api/DungeonsApi.java
@@ -1,19 +1,14 @@
 package mod.gottsch.forge.dungeons2.api;
 
 import mod.gottsch.forge.dungeons2.core.decorator.BlockProvider;
-import mod.gottsch.forge.dungeons2.core.decorator.IRoomElementDecorator;
 import mod.gottsch.forge.dungeons2.core.enums.DungeonMotif;
 import mod.gottsch.forge.dungeons2.core.enums.IDungeonMotif;
-import mod.gottsch.forge.dungeons2.core.enums.IRoomElementType;
 import mod.gottsch.forge.dungeons2.core.pattern.IPatternEnum;
 import mod.gottsch.forge.dungeons2.core.registry.BlockProivderRegistry;
-import mod.gottsch.forge.dungeons2.core.registry.DecoratorRegistry;
 import mod.gottsch.forge.dungeons2.core.registry.EnumRegistry;
 import mod.gottsch.forge.dungeons2.core.registry.PatternRegistry;
 import mod.gottsch.forge.gottschcore.enums.IEnum;
 import mod.gottsch.forge.gottschcore.enums.IRarity;
-import mod.gottsch.forge.gottschcore.random.RandomHelper;
-import net.minecraft.util.RandomSource;
 
 import java.util.List;
 import java.util.Optional;
@@ -72,18 +67,6 @@ public class DungeonsApi {
     public static List<IRarity> getPatterns(String key) {
         List<IPatternEnum> enums = PatternRegistry.getValues(key);
         return enums.stream().map(e -> (IRarity)e).collect(Collectors.toList());
-    }
-
-    public static List<IRoomElementDecorator> getDecorators(IRoomElementType type) {
-        return DecoratorRegistry.get(type);
-    }
-
-    public static Optional<IRoomElementDecorator> getDecorator(RandomSource random, IRoomElementType type) {
-        List<IRoomElementDecorator> list = getDecorators(type);
-        if (list.isEmpty()) {
-            return Optional.empty();
-        }
-        return Optional.of(list.get(RandomHelper.randomInt(random, 0, list.size() - 1)));
     }
 
     public static Optional<BlockProvider> getBlockProvider(IDungeonMotif motif) {